+++ Common/Unity.2D.Entities.Tiny/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Unity.2D.Entities.Tiny.Tests")]
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteRenderer.Tiny")]
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteMask.Tiny")]