+++ SpriteRenderer/Unity.2D.Entities.SpriteRenderer.Authoring.Tiny/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteRenderer.Authoring.Tiny.Editor.Tests")]