+++ SpriteRenderer/Unity.2D.Entities.SpriteRenderer.Authoring.Tiny/shadersrc~/runbee.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-mono ../../../External/dots/TinySamples/Packages/com.unity.dots.runtime/bee~/bee.exe $*
-