+++ SpriteRenderer/Unity.2D.Entities.SpriteRenderer.Authoring/ConvertSpriteRenderer.cs
@@ -6,9 +6,9 @@ using SpriteRenderer = Unity.U2D.Entities.SpriteRenderer;
 
 namespace Unity.U2D.Conversion
 {
-    // replace this with a declarative/autodetect system that can solve for dependency requirements
-    [ConverterVersion("2d", 1)]
+    [ConverterVersion("2d", 3)]
     [UpdateInGroup(typeof(GameObjectDeclareReferencedObjectsGroup))]
+    [WorldSystemFilter(WorldSystemFilterFlags.HybridGameObjectConversion)]
     internal class SpriteRendererDeclareAssets : GameObjectConversionSystem
     {
         protected override void OnUpdate()
@@ -23,6 +23,7 @@ namespace Unity.U2D.Conversion
     
     [ConverterVersion("2d", 3)]
     [UpdateInGroup(typeof(GameObjectConversionGroup))]
+    [WorldSystemFilter(WorldSystemFilterFlags.HybridGameObjectConversion)]
     internal class SpriteRendererConversion : GameObjectConversionSystem
     {
         protected override void OnUpdate()
@@ -40,10 +41,12 @@ namespace Unity.U2D.Conversion
                     Center = uWorldToLocalMatrix.MultiplyPoint(uWorldBounds.center),
                     Extents = uSpriteRenderer.bounds.extents
                 };
-                var uSortingLayerId = uSpriteRenderer.sortingLayerID;
                 
+                var uSortingLayerId = uSpriteRenderer.sortingLayerID;
+                var renderingLayer = uSpriteRenderer.gameObject.layer;
                 DstEntityManager.AddComponentData(entity, new Renderer2D()
                 {
+                    RenderingLayer = renderingLayer,
                     SortingLayer = (short) UnityEngine.SortingLayer.GetLayerValueFromID(uSortingLayerId),
                     OrderInLayer = (short) uSpriteRenderer.sortingOrder,
                     Bounds = localBounds,