+++ SpriteRenderer/Unity.2D.Entities.SpriteRenderer.Tiny/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteRenderer.Authoring.Tiny")]
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteRenderer.Authoring.Tiny.Editor.Tests")]
-
-[assembly: InternalsVisibleTo("Unity.2D.Entities.SpriteMask.Tiny")]