+++ Unity.2D.Entities.Physics.Authoring/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Unity.2D.Entities.Physics.Editor.Tests")]
-