+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.31.0-preview.3] - 2020-10-21
+###Added
+- Added Entities 0.16 support
+
+###Changed
+- 2D runtime for Project Tiny has been moved out of 2D Entities and into the Tiny package
+- 2D Physics has been moved into its own package, 2D Entities Physics
+
 ## [0.29.0-preview.3] - 2020-08-24
 ###Added
 - Added Tiny 0.29 support